createGrid() in Board.h
