+++ PA9_Group2/Board.h
@@ -7,4 +7,4 @@ using namespace std;
 //class for game board with grid, display options, enemy marching path, rules for acceptable builds (can't build on the same square 2x, etc)
 
 
-
+sf::VertexArray createGrid(double width, double height, double cellSize);
