+++ PA9_Group2/PA9_Group2.vcxproj
@@ -158,14 +158,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Grid.cpp" />
+    <ClCompile Include="Board.cpp" />
+    <ClCompile Include="Enemy.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="PA9.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Board.h" />
     <ClInclude Include="Enemy.h" />
-    <ClInclude Include="Grid.h" />
     <ClInclude Include="PA9.h" />
     <ClInclude Include="Player.h" />
     <ClInclude Include="Tower.h" />
